feat(ka): v1.4 QE readiness audit — phases 2–5
Implements the full v1.4 gap remediation plan for the kubernaut-agent
shadow/alignment subsystem across four phases:

Phase 2 — Mechanical Code Fixes:
- Handle all json.Marshal errors across 19+ call sites (GAP-E2)
- Replace fmt.Fprintf(os.Stderr) with structured logr logging (GAP-E3)
- Expand Config.Validate() with bounds checks for 11 fields (GAP-C3)
- Wire ValidateHeaderSources into LLMRuntime validation (GAP-C4)
- Set explicit HTTP client timeouts for DS and LLM (GAP-C5)
- Return structured 500 errors from session lookup (GAP-C6)
- Replace time.Sleep with Eventually/Consistently in tests (GAP-T4)
- Extract marshalAuditResponseData helper (GAP-A3)
- Enable --race flag in unit/integration test targets (QE-1)
- Fix coverage script for Go service recognition (QE-3)
- Add EXPOSE 8081 9090 to Dockerfile (OPS-4)

Phase 3 — Security Hardening:
- Convert constructors/boundary.Generate from panic to error (GAP-E4)
- Cap SanitizeExplanation pre-allocation + early-exit (GAP-S4)
- Add http.MaxBytesReader (1 MiB) on API routes (SEC-1)
- Set Read/Write/IdleTimeout on all 3 HTTP servers (SEC-2)
- Sanitize error details in 500 responses (SEC-3)
- Implement token-bucket rate limiter middleware (SEC-5)
- Add MaxConcurrentInvestigations semaphore (SEC-6)
- Cap Prometheus response body reads via LimitReader (SEC-7)

Phase 4 — Design & Contract Changes:
- Move /config to health port, add /admin/loglevel (SEC-4, OPS-3)
- Wire DrainAndWait into shutdown sequence (GAP-S5)
- Use RemediationID as alignment correlation ID (GAP-A4)
- Promote 4 string-literal EventActions to constants (GAP-T3)
- Embed OpenAPI spec via go:embed + serve /openapi.json (OPS-1)
- Re-resolve credentials on LLM hot-reload (OPS-6)
- Update openapi.json paths /health→/healthz, /ready→/readyz (PROD-1)
- Add typed AnalyzeAccepted, SessionStatus, InvestigationFailed schemas (PROD-2)
- Return 200 with failure body for failed sessions (PROD-3)

Phase 5 — Architecture:
- Split investigator.go (1740→911 lines) into 4 files (ARCH-1)
- Replace interface{} with *InvestigationResult in session APIs (ARCH-3)
- Add audit verbosity config (full/standard/minimal) (GAP-S6)
- Add ActorID/ActorType propagation in audit events (COMP-2)

All 25 unit test suites pass with --race enabled.
