+++ internal/kubernautagent/alignment/investigator_wrapper.go
@@ -55,13 +55,13 @@ type InvestigatorWrapperConfig struct {
 }
 
 // NewInvestigatorWrapper creates an InvestigatorWrapper.
-// Panics if Inner or Evaluator is nil to prevent nil deref during Investigate.
-func NewInvestigatorWrapper(cfg InvestigatorWrapperConfig) *InvestigatorWrapper {
+// Returns an error if Inner or Evaluator is nil to prevent nil deref during Investigate.
+func NewInvestigatorWrapper(cfg InvestigatorWrapperConfig) (*InvestigatorWrapper, error) {
 	if cfg.Inner == nil {
-		panic("alignment.NewInvestigatorWrapper: Inner must not be nil")
+		return nil, fmt.Errorf("alignment.NewInvestigatorWrapper: Inner must not be nil")
 	}
 	if cfg.Evaluator == nil {
-		panic("alignment.NewInvestigatorWrapper: Evaluator must not be nil")
+		return nil, fmt.Errorf("alignment.NewInvestigatorWrapper: Evaluator must not be nil")
 	}
 	logger := cfg.Logger
 	var zero logr.Logger
@@ -80,7 +80,7 @@ func NewInvestigatorWrapper(cfg InvestigatorWrapperConfig) *InvestigatorWrapper 
 		logger:                logger,
 		mode:                  mode,
 		canaryForceEscalation: cfg.CanaryForceEscalation,
-	}
+	}, nil
 }
 
 // Investigate runs a canary integrity check, creates a per-request Observer,
@@ -113,7 +113,10 @@ func (w *InvestigatorWrapper) Investigate(ctx context.Context, signal katypes.Si
 		)
 	}
 
-	observer := NewObserver(w.evaluator)
+	observer, obsErr := NewObserver(w.evaluator)
+	if obsErr != nil {
+		return nil, fmt.Errorf("alignment observer: %w", obsErr)
+	}
 	ctx = WithObserver(ctx, observer)
 
 	if signalContent := BuildSignalInputContent(signal); signalContent != "" {
@@ -194,9 +197,14 @@ func (w *InvestigatorWrapper) emitAlignmentAudit(ctx context.Context, signal kat
 		return
 	}
 
+	correlationID := signal.RemediationID
+	if correlationID == "" {
+		correlationID = signal.Name
+	}
+
 	for _, obs := range verdict.Observations {
 		if obs.Suspicious {
-			event := audit.NewEvent(audit.EventTypeAlignmentStep, signal.Name)
+			event := audit.NewEvent(audit.EventTypeAlignmentStep, correlationID)
 			event.EventAction = audit.ActionAlignmentEvaluate
 			event.EventOutcome = audit.OutcomeFailure
 			event.Data["step_index"] = obs.Step.Index
@@ -207,7 +215,7 @@ func (w *InvestigatorWrapper) emitAlignmentAudit(ctx context.Context, signal kat
 		}
 	}
 
-	event := audit.NewEvent(audit.EventTypeAlignmentVerdict, signal.Name)
+	event := audit.NewEvent(audit.EventTypeAlignmentVerdict, correlationID)
 	event.EventAction = audit.ActionAlignmentVerdict
 	if verdict.Result == VerdictSuspicious {
 		event.EventOutcome = audit.OutcomeFailure
