+++ internal/kubernautagent/investigator/investigator.go
@@ -20,15 +20,11 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
-	"strings"
 	"time"
 
 	"github.com/go-logr/logr"
 	"golang.org/x/sync/errgroup"
-	"k8s.io/apimachinery/pkg/api/meta"
-	"k8s.io/apimachinery/pkg/runtime/schema"
 
-	"github.com/jordigilh/kubernaut/internal/kubernautagent/alignment"
 	"github.com/jordigilh/kubernaut/internal/kubernautagent/audit"
 	"github.com/jordigilh/kubernaut/internal/kubernautagent/enrichment"
 	"github.com/jordigilh/kubernaut/internal/kubernautagent/parser"
@@ -518,116 +514,6 @@ CRITICAL: root_cause_analysis must be a JSON object, NOT a string. Do NOT wrap i
 // re-evaluate its target. If the retry also returns the same kind, the result
 // is accepted as-is (the LLM explicitly confirmed its choice).
 // Issue #847 / DD-HAPI-847, Layer 3: Programmatic Sentinel Validation Gate.
-func (inv *Investigator) sameKindValidationGate(
-	ctx context.Context,
-	result *katypes.InvestigationResult,
-	signal katypes.SignalContext,
-	history []llm.Message,
-	tokens *TokenAccumulator,
-	correlationID string,
-	client llm.Client,
-	modelName string,
-	runtimeParams llm.RuntimeParams,
-) *katypes.InvestigationResult {
-	if signal.ResourceKind == "" || result.RemediationTarget.Kind == "" {
-		return result
-	}
-	if !strings.EqualFold(result.RemediationTarget.Kind, signal.ResourceKind) {
-		return result
-	}
-
-	inv.logger.Info("same-kind validation gate triggered: remediation_target.kind matches signal resource_kind",
-		"target_kind", result.RemediationTarget.Kind,
-		"signal_resource_kind", signal.ResourceKind,
-		"correlation_id", correlationID)
-
-	gateEvent := audit.NewEvent(audit.EventTypeLLMRequest, correlationID)
-	gateEvent.EventAction = "same_kind_validation_gate"
-	gateEvent.EventOutcome = audit.OutcomeSuccess
-	gateEvent.Data["model"] = modelName
-	gateEvent.Data["prompt_length"] = 0
-	gateEvent.Data["prompt_preview"] = ""
-	gateEvent.Data["signal_resource_kind"] = signal.ResourceKind
-	gateEvent.Data["target_kind"] = result.RemediationTarget.Kind
-	gateEvent.Data["target_name"] = result.RemediationTarget.Name
-	audit.StoreBestEffort(ctx, inv.auditStore, gateEvent, inv.auditLog())
-
-	correctionMsg := fmt.Sprintf(
-		`Your remediation_target.kind is "%s", which is the same resource kind as the input signal. `+
-			`Signals often propagate upward: workload-level issues manifest as conditions on parent resources `+
-			`(e.g., pod memory leaks cause node DiskPressure, deployment misconfigurations appear as node conditions). `+
-			`Please re-evaluate: is a child resource (Deployment, StatefulSet, DaemonSet, Pod) the actual root cause `+
-			`whose configuration should be modified? If after re-evaluation you are confident the %s itself is the `+
-			`correct remediation target, confirm by resubmitting with the same target and explain why in your `+
-			`due_diligence.target_accuracy field.`,
-		result.RemediationTarget.Kind,
-		result.RemediationTarget.Kind,
-	)
-
-	submitOnlyTools := []llm.ToolDefinition{
-		{
-			Name:        SubmitResultToolName,
-			Description: "Submit root cause analysis result.",
-			Parameters:  parser.RCAResultSchema(),
-		},
-	}
-
-	retryMessages := make([]llm.Message, len(history))
-	copy(retryMessages, history)
-	retryMessages = append(retryMessages,
-		llm.Message{Role: "user", Content: correctionMsg},
-	)
-
-	resp, err := llm.ChatWithParams(ctx, client, llm.ChatRequest{
-		Messages: retryMessages,
-		Tools:    submitOnlyTools,
-		Options:  llm.ChatOptions{JSONMode: true, OutputSchema: parser.RCAResultSchema()},
-	}, runtimeParams)
-	if err != nil {
-		inv.logger.Error(err, "same-kind validation gate retry failed, keeping original result",
-			"correlation_id", correlationID)
-		return result
-	}
-	if tokens != nil {
-		tokens.Add(resp.Usage)
-	}
-
-	var retryContent string
-	for _, tc := range resp.ToolCalls {
-		if tc.Name == SubmitResultToolName {
-			retryContent = tc.Arguments
-			break
-		}
-	}
-	if retryContent == "" && resp.Message.Content != "" {
-		retryContent = resp.Message.Content
-	}
-	if retryContent == "" {
-		inv.logger.Info("same-kind validation gate: no content in retry response, keeping original",
-			"correlation_id", correlationID)
-		return result
-	}
-
-	retryResult, parseErr := inv.resultParser.Parse(retryContent)
-	if parseErr != nil {
-		inv.logger.Error(parseErr, "same-kind validation gate: retry parse failed, keeping original",
-			"correlation_id", correlationID)
-		return result
-	}
-
-	if retryResult.RemediationTarget.Kind == "" && result.RemediationTarget.Kind != "" {
-		inv.logger.Info("same-kind validation gate: retry lost remediation_target, keeping original",
-			"original_target", result.RemediationTarget.Kind+"/"+result.RemediationTarget.Name,
-			"correlation_id", correlationID)
-		return result
-	}
-
-	inv.logger.Info("same-kind validation gate: accepted retry result",
-		"original_target", result.RemediationTarget.Kind+"/"+result.RemediationTarget.Name,
-		"retry_target", retryResult.RemediationTarget.Kind+"/"+retryResult.RemediationTarget.Name,
-		"correlation_id", correlationID)
-	return retryResult
-}
 
 func (inv *Investigator) runWorkflowSelection(ctx context.Context, signal katypes.SignalContext, rcaSummary string, enrichData *prompt.EnrichmentData, p1Ctx *prompt.Phase1Data, tokens *TokenAccumulator, correlationID string, client llm.Client, modelName string, runtimeParams llm.RuntimeParams) (*katypes.InvestigationResult, error) {
 	// Attach signal context so workflow discovery tools (list_available_actions,
@@ -894,67 +780,6 @@ Do NOT respond with plain text. You MUST call one of the above tools.`
 // into the InvestigationResult, replicating KA's behavior of including
 // execution_engine, execution_bundle, execution_bundle_digest, and
 // service_account_name so downstream controllers (WE) can use them.
-func enrichFromCatalog(result *katypes.InvestigationResult, v *parser.Validator) {
-	if result == nil || v == nil || result.WorkflowID == "" {
-		return
-	}
-	meta, ok := v.GetWorkflowMeta(result.WorkflowID)
-	if !ok {
-		return
-	}
-	if result.ExecutionEngine == "" {
-		result.ExecutionEngine = meta.ExecutionEngine
-	}
-	if result.ExecutionBundle == "" {
-		result.ExecutionBundle = meta.ExecutionBundle
-	}
-	if result.ExecutionBundleDigest == "" {
-		result.ExecutionBundleDigest = meta.ExecutionBundleDigest
-	}
-	if result.ServiceAccountName == "" {
-		result.ServiceAccountName = meta.ServiceAccountName
-	}
-	result.WorkflowVersion = meta.Version
-}
-
-// CheckWorkflowTargetAlignment verifies that the selected workflow's component
-// scope includes the RCA remediation target kind. Emits an audit event and
-// appends a warning on mismatch. This is a WARNING-level gate: misalignment
-// does not force human review because the LLM may have legitimate cross-type
-// reasoning (Issue #934).
-func CheckWorkflowTargetAlignment(ctx context.Context, result *katypes.InvestigationResult, v *parser.Validator, correlationID string, auditStore audit.AuditStore, logger logr.Logger) {
-	if result == nil || v == nil || result.WorkflowID == "" {
-		return
-	}
-	meta, ok := v.GetWorkflowMeta(result.WorkflowID)
-	if !ok {
-		return
-	}
-
-	aligned := meta.MatchesTargetKind(result.RemediationTarget.Kind)
-
-	ev := audit.NewEvent(audit.EventTypeLLMRequest, correlationID)
-	ev.EventAction = "workflow_target_alignment_gate"
-	if aligned {
-		ev.EventOutcome = audit.OutcomeSuccess
-	} else {
-		ev.EventOutcome = audit.OutcomeFailure
-	}
-	ev.Data["model"] = ""
-	ev.Data["prompt_length"] = 0
-	ev.Data["prompt_preview"] = ""
-	ev.Data["workflow_id"] = result.WorkflowID
-	ev.Data["target_kind"] = result.RemediationTarget.Kind
-	ev.Data["workflow_components"] = meta.Component
-	ev.Data["aligned"] = aligned
-	audit.StoreBestEffort(ctx, auditStore, ev, logger)
-
-	if !aligned {
-		result.Warnings = append(result.Warnings,
-			fmt.Sprintf("Workflow %q target kind %q is not in the workflow's component scope %v",
-				result.WorkflowID, result.RemediationTarget.Kind, meta.Component))
-	}
-}
 
 // runLLMLoop executes the multi-turn LLM interaction loop with tool
 // execution routed through the registry. correlationID is propagated to
@@ -1068,7 +893,7 @@ func (inv *Investigator) runLLMLoop(ctx context.Context, messages []llm.Message,
 				"correlation_id", correlationID)
 
 			truncEvent := audit.NewEvent(audit.EventTypeLLMResponse, correlationID)
-			truncEvent.EventAction = "truncation_detected"
+			truncEvent.EventAction = audit.ActionTruncationDetected
 			truncEvent.EventOutcome = audit.OutcomeFailure
 			truncEvent.Data["has_analysis"] = resp.Message.Content != ""
 			truncEvent.Data["analysis_length"] = len(resp.Message.Content)
@@ -1092,649 +917,5 @@ func (inv *Investigator) runLLMLoop(ctx context.Context, messages []llm.Message,
 	return &ExhaustedResult{Reason: "max turns exhausted"}, nil
 }
 
-// escalateMaxTokens computes a higher MaxTokens value for truncation recovery.
-// If the truncated response used N completion tokens, we request 2x. Falls back
-// to a default of 8192 if the usage data is unavailable.
-func escalateMaxTokens(completionTokens int) int {
-	if completionTokens > 0 {
-		escalated := completionTokens * 2
-		if escalated > 16384 {
-			return 16384
-		}
-		return escalated
-	}
-	return 8192
-}
 
-func totalPromptLength(messages []llm.Message) int {
-	total := 0
-	for _, m := range messages {
-		total += len(m.Content)
-	}
-	return total
-}
 
-func lastUserMessage(messages []llm.Message, maxLen int) string {
-	for i := len(messages) - 1; i >= 0; i-- {
-		if messages[i].Role == "user" {
-			return truncatePreview(messages[i].Content, maxLen)
-		}
-	}
-	return ""
-}
-
-func truncatePreview(s string, maxLen int) string {
-	if len(s) <= maxLen {
-		return s
-	}
-	return s[:maxLen]
-}
-
-func messagesToAuditFormat(messages []llm.Message) []map[string]interface{} {
-	out := make([]map[string]interface{}, len(messages))
-	for i, m := range messages {
-		entry := map[string]interface{}{
-			"role":    m.Role,
-			"content": m.Content,
-		}
-		if m.ToolCallID != "" {
-			entry["tool_call_id"] = m.ToolCallID
-		}
-		if m.ToolName != "" {
-			entry["name"] = m.ToolName
-		}
-		out[i] = entry
-	}
-	return out
-}
-
-func toolNames(defs []llm.ToolDefinition) []string {
-	names := make([]string, len(defs))
-	for i, d := range defs {
-		names[i] = d.Name
-	}
-	return names
-}
-
-func (inv *Investigator) toolDefinitionsForPhase(phase katypes.Phase) []llm.ToolDefinition {
-	var defs []llm.ToolDefinition
-	if inv.registry != nil {
-		phaseTools := inv.registry.ToolsForPhase(phase, inv.phaseTools)
-		defs = make([]llm.ToolDefinition, 0, len(phaseTools)+2)
-		for _, t := range phaseTools {
-			defs = append(defs, llm.ToolDefinition{
-				Name:        t.Name(),
-				Description: t.Description(),
-				Parameters:  t.Parameters(),
-			})
-		}
-	}
-
-	if phase == katypes.PhaseWorkflowDiscovery {
-		defs = append(defs,
-			llm.ToolDefinition{
-				Name:        SubmitResultWithWorkflowToolName,
-				Description: "Submit investigation result WITH a selected workflow. Call this when you have identified a matching workflow.",
-				Parameters:  parser.WithWorkflowResultSchema(),
-			},
-			llm.ToolDefinition{
-				Name:        SubmitResultNoWorkflowToolName,
-				Description: "Submit investigation result when NO matching workflow exists. Call this when none of the available workflows can remediate the incident.",
-				Parameters:  parser.NoWorkflowResultSchema(),
-			},
-		)
-	} else {
-		defs = append(defs, llm.ToolDefinition{
-			Name:        SubmitResultToolName,
-			Description: "Submit the final investigation result as structured JSON. Call this tool when your analysis is complete.",
-			Parameters:  submitResultSchemaForPhase(phase),
-		})
-	}
-	return defs
-}
-
-// submitResultSchemaForPhase returns the phase-appropriate JSON Schema for
-// the submit_result tool. RCA uses a restricted schema without workflow/escalation
-// fields; workflow discovery uses the full InvestigationResultSchema.
-func submitResultSchemaForPhase(phase katypes.Phase) json.RawMessage {
-	if phase == katypes.PhaseRCA {
-		return parser.RCAResultSchema()
-	}
-	return parser.InvestigationResultSchema()
-}
-
-func (inv *Investigator) executeTool(ctx context.Context, name string, args json.RawMessage) string {
-	if inv.registry == nil {
-		return toolErrorJSON("no registry configured for tool " + name)
-	}
-
-	if ar := inv.pipeline.AnomalyDetector.CheckToolCall(name, args); !ar.Allowed {
-		inv.logger.Info("anomaly detector rejected tool call",
-			"tool", name,
-			"reason", ar.Reason,
-		)
-		return toolErrorJSON(ar.Reason)
-	}
-
-	result, err := inv.registry.Execute(ctx, name, args)
-	if err != nil {
-		inv.logger.Error(err, "tool execution failed",
-			"tool", name,
-		)
-		if ar := inv.pipeline.AnomalyDetector.RecordFailure(name, args); !ar.Allowed {
-			errResult := toolErrorJSON(ar.Reason)
-			alignment.SubmitToolStep(ctx, name, errResult)
-			return errResult
-		}
-		errResult := toolErrorJSON(err.Error())
-		alignment.SubmitToolStep(ctx, name, errResult)
-		return errResult
-	}
-
-	if inv.pipeline.Sanitizer != nil {
-		sanitized, sanitizeErr := inv.pipeline.Sanitizer.Run(ctx, result)
-		if sanitizeErr != nil {
-			inv.logger.Error(sanitizeErr, "sanitization failed, returning raw output",
-				"tool", name,
-			)
-		} else {
-			result = sanitized
-		}
-	}
-
-	if inv.pipeline.Summarizer != nil {
-		summarized, sumErr := inv.pipeline.Summarizer.MaybeSummarize(ctx, name, result)
-		if sumErr != nil {
-			inv.logger.Error(sumErr, "summarization failed, returning unsummarized output",
-				"tool", name,
-			)
-		} else {
-			result = summarized
-		}
-	}
-
-	if inv.pipeline.MaxToolOutputSize > 0 {
-		result = summarizer.TruncateToolOutput(result, name, inv.pipeline.MaxToolOutputSize)
-	}
-
-	alignment.SubmitToolStep(ctx, name, result)
-
-	return result
-}
-
-func (inv *Investigator) emitResponseComplete(ctx context.Context, result *katypes.InvestigationResult, tokens *TokenAccumulator, correlationID string) {
-	completeEvent := audit.NewEvent(audit.EventTypeResponseComplete, correlationID)
-	completeEvent.EventAction = audit.ActionResponseSent
-	completeEvent.EventOutcome = audit.OutcomeSuccess
-	for k, v := range tokens.AuditData() {
-		completeEvent.Data[k] = v
-	}
-	if b, err := json.Marshal(ResultToAuditJSON(result)); err == nil {
-		completeEvent.Data["response_data"] = string(b)
-	}
-	audit.StoreBestEffort(ctx, inv.auditStore, completeEvent, inv.auditLog())
-}
-
-func (inv *Investigator) emitRCAComplete(ctx context.Context, result *katypes.InvestigationResult, tokens *TokenAccumulator, correlationID string) {
-	ev := audit.NewEvent(audit.EventTypeRCAComplete, correlationID)
-	ev.EventAction = audit.ActionLLMResponse
-	ev.EventOutcome = audit.OutcomeSuccess
-	for k, v := range tokens.AuditData() {
-		ev.Data[k] = v
-	}
-	if b, err := json.Marshal(ResultToAuditJSON(result)); err == nil {
-		ev.Data["response_data"] = string(b)
-	}
-	audit.StoreBestEffort(ctx, inv.auditStore, ev, inv.auditLog())
-}
-
-func ResultToAuditJSON(r *katypes.InvestigationResult) map[string]interface{} {
-	m := map[string]interface{}{
-		"rca_summary":        r.RCASummary,
-		"severity":           r.Severity,
-		"confidence":         r.Confidence,
-		"needs_human_review": r.HumanReviewNeeded,
-	}
-	if r.WorkflowID != "" {
-		m["workflow_id"] = r.WorkflowID
-	}
-	if r.ExecutionBundle != "" {
-		m["execution_bundle"] = r.ExecutionBundle
-	}
-	if r.ExecutionBundleDigest != "" {
-		m["execution_bundle_digest"] = r.ExecutionBundleDigest
-	}
-	if r.ExecutionEngine != "" {
-		m["execution_engine"] = r.ExecutionEngine
-	}
-	if len(r.ContributingFactors) > 0 {
-		m["contributing_factors"] = r.ContributingFactors
-	}
-	if r.HumanReviewReason != "" {
-		m["human_review_reason"] = r.HumanReviewReason
-	} else if r.Reason != "" {
-		m["human_review_reason"] = r.Reason
-	}
-	if r.Reason != "" {
-		m["reason"] = r.Reason
-	}
-	if r.IsActionable != nil {
-		m["is_actionable"] = *r.IsActionable
-	}
-	if r.SignalName != "" {
-		m["signal_name"] = r.SignalName
-	}
-	if r.DetectedLabels != nil {
-		m["detected_labels"] = r.DetectedLabels
-	}
-	if len(r.ValidationAttemptsHistory) > 0 {
-		m["validation_attempts_history"] = r.ValidationAttemptsHistory
-	}
-	if len(r.Warnings) > 0 {
-		m["warnings"] = r.Warnings
-	}
-	if len(r.Parameters) > 0 {
-		m["parameters"] = r.Parameters
-	}
-	if r.RemediationTarget.Kind != "" {
-		m["remediation_target"] = map[string]interface{}{
-			"kind":      r.RemediationTarget.Kind,
-			"name":      r.RemediationTarget.Name,
-			"namespace": r.RemediationTarget.Namespace,
-		}
-	}
-	if len(r.AlternativeWorkflows) > 0 {
-		alts := make([]map[string]interface{}, len(r.AlternativeWorkflows))
-		for i, alt := range r.AlternativeWorkflows {
-			a := map[string]interface{}{
-				"workflow_id": alt.WorkflowID,
-				"confidence":  alt.Confidence,
-			}
-			if alt.Rationale != "" {
-				a["rationale"] = alt.Rationale
-			}
-			if alt.ExecutionBundle != "" {
-				a["execution_bundle"] = alt.ExecutionBundle
-			}
-			alts[i] = a
-		}
-		m["alternative_workflows"] = alts
-	}
-	if len(r.CausalChain) > 0 {
-		m["causal_chain"] = r.CausalChain
-	}
-	if r.DueDiligence != nil {
-		m["due_diligence"] = map[string]interface{}{
-			"causal_completeness":    r.DueDiligence.CausalCompleteness,
-			"target_accuracy":        r.DueDiligence.TargetAccuracy,
-			"evidence_sufficiency":   r.DueDiligence.EvidenceSufficiency,
-			"alternative_hypotheses": r.DueDiligence.AlternativeHypotheses,
-			"scope_completeness":     r.DueDiligence.ScopeCompleteness,
-			"proportionality":        r.DueDiligence.Proportionality,
-			"regression_awareness":   r.DueDiligence.RegressionAwareness,
-			"confidence_calibration": r.DueDiligence.ConfidenceCalibration,
-		}
-	}
-	return m
-}
-
-func (inv *Investigator) emitValidationEvent(ctx context.Context, attempt, maxAttempts int, isValid bool, errors []string, workflowID, correlationID string) {
-	valEvent := audit.NewEvent(audit.EventTypeValidationAttempt, correlationID)
-	valEvent.EventAction = audit.ActionValidation
-	if isValid {
-		valEvent.EventOutcome = audit.OutcomeSuccess
-	} else {
-		valEvent.EventOutcome = audit.OutcomeFailure
-	}
-	valEvent.Data["attempt"] = attempt
-	valEvent.Data["max_attempts"] = maxAttempts
-	valEvent.Data["is_valid"] = isValid
-	valEvent.Data["errors"] = errors
-	valEvent.Data["workflow_id"] = workflowID
-	valEvent.Data["is_final_attempt"] = attempt == maxAttempts
-	audit.StoreBestEffort(ctx, inv.auditStore, valEvent, inv.auditLog())
-}
-
-func toolErrorJSON(msg string) string {
-	payload := struct {
-		Status string `json:"status"`
-		Error  string `json:"error"`
-	}{Status: "error", Error: msg}
-	b, _ := json.Marshal(payload)
-	return string(b)
-}
-
-// BuildPhase1Context extracts structured assessment fields from the Phase 1
-// InvestigationResult for propagation into Phase 3 (HAPI parity: #715).
-func BuildPhase1Context(rcaResult *katypes.InvestigationResult) *prompt.Phase1Data {
-	if rcaResult == nil {
-		return nil
-	}
-	return &prompt.Phase1Data{
-		Severity:            rcaResult.Severity,
-		ContributingFactors: rcaResult.ContributingFactors,
-		RemediationTarget: prompt.Phase1RemediationTarget{
-			Kind:      rcaResult.RemediationTarget.Kind,
-			Name:      rcaResult.RemediationTarget.Name,
-			Namespace: rcaResult.RemediationTarget.Namespace,
-		},
-		InvestigationOutcome:  rcaResult.InvestigationOutcome,
-		Confidence:            rcaResult.Confidence,
-		InvestigationAnalysis: rcaResult.InvestigationAnalysis,
-		CausalChain:           rcaResult.CausalChain,
-		DueDiligence:          rcaResult.DueDiligence,
-	}
-}
-
-// MergePhase1Fallbacks applies Phase 1 assessment fields to the Phase 3 result
-// when Phase 3 did not produce them. Matches HAPI's result.setdefault() pattern:
-// Phase 3 values always take precedence; Phase 1 fills in gaps only.
-func MergePhase1Fallbacks(result *katypes.InvestigationResult, p1 *prompt.Phase1Data) {
-	if result == nil || p1 == nil {
-		return
-	}
-	if result.Severity == "" && p1.Severity != "" {
-		result.Severity = p1.Severity
-	}
-	if len(result.ContributingFactors) == 0 && len(p1.ContributingFactors) > 0 {
-		result.ContributingFactors = p1.ContributingFactors
-	}
-	if result.Confidence == 0 && p1.Confidence > 0 {
-		result.Confidence = p1.Confidence
-	}
-	if result.InvestigationOutcome == "" && p1.InvestigationOutcome != "" {
-		result.InvestigationOutcome = p1.InvestigationOutcome
-		parser.ApplyInvestigationOutcome(result, p1.InvestigationOutcome)
-		// #301 defense-in-depth: Phase 1 problem_resolved overrides
-		// contradictory HumanReviewNeeded set by Phase 3 (e.g. the
-		// SubmitNoWorkflowResult branch hardcodes HumanReviewNeeded=true,
-		// but that should not apply when the investigation is resolved).
-		if p1.InvestigationOutcome == "problem_resolved" && result.HumanReviewNeeded {
-			result.HumanReviewNeeded = false
-			result.HumanReviewReason = ""
-		}
-	}
-	if len(result.CausalChain) == 0 && len(p1.CausalChain) > 0 {
-		result.CausalChain = p1.CausalChain
-	}
-	if result.DueDiligence == nil && p1.DueDiligence != nil {
-		result.DueDiligence = p1.DueDiligence
-	}
-}
-
-func signalToPrompt(s katypes.SignalContext) prompt.SignalData {
-	return prompt.SignalData{
-		Name:                       s.Name,
-		Namespace:                  s.Namespace,
-		Severity:                   s.Severity,
-		Message:                    s.Message,
-		ResourceKind:               s.ResourceKind,
-		ResourceName:               s.ResourceName,
-		ClusterName:                s.ClusterName,
-		Environment:                s.Environment,
-		Priority:                   s.Priority,
-		RiskTolerance:              s.RiskTolerance,
-		SignalSource:               s.SignalSource,
-		BusinessCategory:           s.BusinessCategory,
-		Description:                s.Description,
-		SignalMode:                 s.SignalMode,
-		FiringTime:                 s.FiringTime,
-		ReceivedTime:               s.ReceivedTime,
-		IsDuplicate:                s.IsDuplicate,
-		OccurrenceCount:            s.OccurrenceCount,
-		DeduplicationWindowMinutes: s.DeduplicationWindowMinutes,
-		FirstSeen:                  s.FirstSeen,
-		LastSeen:                   s.LastSeen,
-		SignalAnnotations:          s.SignalAnnotations,
-	}
-}
-
-func toPromptEnrichment(data *enrichment.EnrichmentResult) *prompt.EnrichmentData {
-	if data == nil {
-		return nil
-	}
-	pe := &prompt.EnrichmentData{}
-
-	for _, entry := range data.OwnerChain {
-		if entry.Namespace != "" {
-			pe.OwnerChain = append(pe.OwnerChain, entry.Kind+"/"+entry.Name+"("+entry.Namespace+")")
-		} else {
-			pe.OwnerChain = append(pe.OwnerChain, entry.Kind+"/"+entry.Name)
-		}
-	}
-
-	if data.DetectedLabels != nil {
-		pe.DetectedLabels = detectedLabelsToPromptMap(data.DetectedLabels)
-	}
-
-	if len(data.QuotaDetails) > 0 {
-		pe.QuotaDetails = data.QuotaDetails
-	}
-
-	pe.HistoryResult = data.RemediationHistory
-	return pe
-}
-
-func detectedLabelsToPromptMap(dl *enrichment.DetectedLabels) map[string]string {
-	m := make(map[string]string)
-	if dl.GitOpsManaged {
-		m["gitOpsManaged"] = "true"
-		if dl.GitOpsTool != "" {
-			m["gitOpsTool"] = dl.GitOpsTool
-		}
-	}
-	if dl.HPAEnabled {
-		m["hpaEnabled"] = "true"
-	}
-	if dl.PDBProtected {
-		m["pdbProtected"] = "true"
-	}
-	if dl.Stateful {
-		m["stateful"] = "true"
-	}
-	if dl.HelmManaged {
-		m["helmManaged"] = "true"
-	}
-	if dl.NetworkIsolated {
-		m["networkIsolated"] = "true"
-	}
-	if dl.ServiceMesh != "" {
-		m["serviceMesh"] = dl.ServiceMesh
-	}
-	if dl.ResourceQuotaConstrained {
-		m["resourceQuotaConstrained"] = "true"
-	}
-	if len(dl.FailedDetections) > 0 {
-		m["failedDetections"] = strings.Join(dl.FailedDetections, ",")
-	}
-	return m
-}
-
-func detectedLabelsToResult(dl *enrichment.DetectedLabels) map[string]interface{} {
-	m := make(map[string]interface{})
-	m["gitOpsManaged"] = dl.GitOpsManaged
-	if dl.GitOpsTool != "" {
-		m["gitOpsTool"] = dl.GitOpsTool
-	}
-	m["hpaEnabled"] = dl.HPAEnabled
-	m["pdbProtected"] = dl.PDBProtected
-	m["stateful"] = dl.Stateful
-	m["helmManaged"] = dl.HelmManaged
-	m["networkIsolated"] = dl.NetworkIsolated
-	if dl.ServiceMesh != "" {
-		m["serviceMesh"] = dl.ServiceMesh
-	}
-	m["resourceQuotaConstrained"] = dl.ResourceQuotaConstrained
-	if len(dl.FailedDetections) > 0 {
-		m["failedDetections"] = dl.FailedDetections
-	}
-	return m
-}
-
-func attachDetectedLabels(result *katypes.InvestigationResult, enrichData *enrichment.EnrichmentResult) {
-	if result == nil || enrichData == nil || enrichData.DetectedLabels == nil {
-		return
-	}
-	result.DetectedLabels = detectedLabelsToResult(enrichData.DetectedLabels)
-}
-
-// InjectRemediationTarget resolves the authoritative remediation target.
-//
-// Root owner resolution priority:
-//  1. Owner chain last entry (most common: Pod → RS → Deployment)
-//  2. Enrichment source identity (after re-enrichment, the chain is empty
-//     because the enriched resource IS the root — see #694)
-//  3. Signal identity (fallback when enrichData is nil)
-//
-// LLM target handling:
-//   - Kind == "" or same as root: override with K8s-verified root identity
-//   - Different Kind (cross-type): preserve the LLM's target
-func InjectRemediationTarget(result *katypes.InvestigationResult, signal katypes.SignalContext, enrichData *enrichment.EnrichmentResult) {
-	if result == nil {
-		return
-	}
-	rootKind := signal.ResourceKind
-	rootName := signal.ResourceName
-	rootNS := signal.Namespace
-	if enrichData != nil && len(enrichData.OwnerChain) > 0 {
-		root := enrichData.OwnerChain[len(enrichData.OwnerChain)-1]
-		rootKind = root.Kind
-		rootName = root.Name
-		rootNS = root.Namespace
-	} else if enrichData != nil && enrichData.ResourceKind != "" {
-		rootKind = enrichData.ResourceKind
-		rootName = enrichData.ResourceName
-		rootNS = enrichData.ResourceNamespace
-	}
-
-	llmKind := result.RemediationTarget.Kind
-	if llmKind == "" || llmKind == rootKind {
-		result.RemediationTarget = katypes.RemediationTarget{
-			Kind:      rootKind,
-			Name:      rootName,
-			Namespace: rootNS,
-		}
-		return
-	}
-
-	// BR-496 v2 / BR-HAPI-261 AC#5: if the LLM's kind is a descendant
-	// in the ownership hierarchy (e.g. Pod when root is Deployment),
-	// resolve upward to the K8s-verified root owner. Only preserve
-	// the LLM's target when its kind is genuinely cross-type (not in
-	// the owner chain at all, e.g. Node vs Deployment).
-	if enrichData != nil && isKindInOwnerChain(llmKind, signal.ResourceKind, enrichData.OwnerChain) {
-		result.RemediationTarget = katypes.RemediationTarget{
-			Kind:      rootKind,
-			Name:      rootName,
-			Namespace: rootNS,
-		}
-		return
-	}
-}
-
-// isKindInOwnerChain returns true when the given kind matches the signal's own
-// resource kind or any entry in the enrichment owner chain. This identifies
-// descendants within the same K8s ownership hierarchy (e.g. Pod, ReplicaSet
-// under a Deployment) as opposed to genuinely cross-type targets (e.g. Node).
-func isKindInOwnerChain(kind, signalKind string, chain []enrichment.OwnerChainEntry) bool {
-	if strings.EqualFold(kind, signalKind) {
-		return true
-	}
-	for _, entry := range chain {
-		if strings.EqualFold(entry.Kind, kind) {
-			return true
-		}
-	}
-	return false
-}
-
-// injectTargetResourceParameters merges TARGET_RESOURCE_NAME, TARGET_RESOURCE_KIND,
-// and TARGET_RESOURCE_NAMESPACE into result.Parameters from the final
-// RemediationTarget. KA injected these so that WorkflowExecution Jobs receive
-// the correct target identity as environment variables.
-func injectTargetResourceParameters(result *katypes.InvestigationResult) {
-	if result == nil || result.RemediationTarget.Kind == "" {
-		return
-	}
-	if result.Parameters == nil {
-		result.Parameters = make(map[string]interface{})
-	}
-	result.Parameters["TARGET_RESOURCE_NAME"] = result.RemediationTarget.Name
-	result.Parameters["TARGET_RESOURCE_KIND"] = result.RemediationTarget.Kind
-	result.Parameters["TARGET_RESOURCE_NAMESPACE"] = result.RemediationTarget.Namespace
-}
-
-// allLabelDetectionsFailed returns true when every detection category is in
-// FailedDetections, indicating the target resource could not be fetched at all.
-// Used to decide whether to keep the original signal-target labels instead.
-func allLabelDetectionsFailed(labels *enrichment.DetectedLabels) bool {
-	if labels == nil {
-		return false
-	}
-	return len(labels.FailedDetections) >= len(enrichment.AllDetectionCategories)
-}
-
-// EnrichmentCacheKey returns the dedup cache key for a (kind, name, namespace) tuple (#764).
-func EnrichmentCacheKey(kind, name, namespace string) string {
-	return kind + "/" + name + "/" + namespace
-}
-
-// resolveEnrichmentCached wraps resolveEnrichment with a per-call cache (#764).
-// If the same (kind, name, namespace) was already enriched in this investigation,
-// returns the cached result without making another API call.
-func (inv *Investigator) resolveEnrichmentCached(ctx context.Context, cache map[string]*enrichment.EnrichmentResult, kind, name, namespace, incidentID string) *enrichment.EnrichmentResult {
-	key := EnrichmentCacheKey(kind, name, namespace)
-	if cached, ok := cache[key]; ok {
-		inv.logger.Info("enrichment cache hit, reusing cached result",
-			"kind", kind, "name", name, "namespace", namespace)
-		return cached
-	}
-	result := inv.resolveEnrichment(ctx, kind, name, namespace, incidentID)
-	cache[key] = result
-	return result
-}
-
-// normalizeNamespace forces namespace="" for cluster-scoped resources (#763).
-// If no ScopeResolver is configured, returns the namespace unchanged.
-func (inv *Investigator) normalizeNamespace(kind, namespace string) string {
-	if inv.scopeResolver == nil {
-		return namespace
-	}
-	isCluster, err := inv.scopeResolver.IsClusterScoped(kind)
-	if err != nil {
-		inv.logger.Error(err, "ScopeResolver error, preserving namespace",
-			"kind", kind)
-		return namespace
-	}
-	if isCluster {
-		return ""
-	}
-	return namespace
-}
-
-// mapperScopeResolver implements ScopeResolver using a RESTMapper.
-type mapperScopeResolver struct {
-	mapper meta.RESTMapper
-}
-
-// NewMapperScopeResolver creates a ScopeResolver backed by a RESTMapper.
-func NewMapperScopeResolver(mapper meta.RESTMapper) ScopeResolver {
-	return &mapperScopeResolver{mapper: mapper}
-}
-
-func (r *mapperScopeResolver) IsClusterScoped(kind string) (bool, error) {
-	plural := strings.ToLower(kind) + "s"
-	gvr, err := r.mapper.ResourceFor(schema.GroupVersionResource{Resource: plural})
-	if err != nil {
-		return false, err
-	}
-	gvk, err := r.mapper.KindFor(gvr)
-	if err != nil {
-		return false, err
-	}
-	mapping, err := r.mapper.RESTMapping(gvk.GroupKind(), gvk.Version)
-	if err != nil {
-		return false, err
-	}
-	return mapping.Scope.Name() != meta.RESTScopeNameNamespace, nil
-}
