+++ pkg/shared/health/server.go
@@ -38,5 +38,7 @@ func NewHealthServer(addr string, liveness, readiness http.HandlerFunc) *http.Se
 		Addr:              addr,
 		Handler:           mux,
 		ReadHeaderTimeout: 5 * time.Second,
+		ReadTimeout:       5 * time.Second,
+		WriteTimeout:      10 * time.Second,
 	}
 }
