+++ test/unit/kubernautagent/session/metadata_test.go
@@ -25,6 +25,7 @@ import (
 
 	"github.com/go-logr/logr"
 	"github.com/jordigilh/kubernaut/internal/kubernautagent/session"
+	katypes "github.com/jordigilh/kubernaut/internal/kubernautagent/types"
 )
 
 var _ = Describe("Session Metadata — #433", func() {
@@ -38,9 +39,9 @@ var _ = Describe("Session Metadata — #433", func() {
 				"incident_id": "e2e-ka-001-oom",
 			}
 
-			id, err := manager.StartInvestigation(context.Background(), func(ctx context.Context) (interface{}, error) {
-				time.Sleep(50 * time.Millisecond)
-				return "done", nil
+			id, err := manager.StartInvestigation(context.Background(), func(ctx context.Context) (*katypes.InvestigationResult, error) {
+				time.Sleep(50 * time.Millisecond) // APPROVED EXCEPTION: simulated investigation work
+				return &katypes.InvestigationResult{RCASummary: "done"}, nil
 			}, metadata)
 			Expect(err).NotTo(HaveOccurred())
 			Expect(id).NotTo(BeEmpty())
@@ -61,8 +62,8 @@ var _ = Describe("Session Metadata — #433", func() {
 				"incident_id": "e2e-ka-002-crash",
 			}
 
-			id, err := manager.StartInvestigation(context.Background(), func(ctx context.Context) (interface{}, error) {
-				return "result", nil
+			id, err := manager.StartInvestigation(context.Background(), func(ctx context.Context) (*katypes.InvestigationResult, error) {
+				return &katypes.InvestigationResult{RCASummary: "result"}, nil
 			}, metadata)
 			Expect(err).NotTo(HaveOccurred())
 
@@ -86,8 +87,8 @@ var _ = Describe("Session Metadata — #433", func() {
 			store := session.NewStore(5 * time.Minute)
 			manager := session.NewManager(store, logr.Discard())
 
-			id, err := manager.StartInvestigation(context.Background(), func(ctx context.Context) (interface{}, error) {
-				return "result", nil
+			id, err := manager.StartInvestigation(context.Background(), func(ctx context.Context) (*katypes.InvestigationResult, error) {
+				return &katypes.InvestigationResult{RCASummary: "result"}, nil
 			}, nil)
 			Expect(err).NotTo(HaveOccurred())
